Move files to examples + split examples apart
